Fix RiotTests.
